bash script for running from terminal
This bash script takes a diagram files' name (without extension). Then it makes a copy of the DiagWand.html named after the diagram so that the new file knows the path of the diagram file. Then it opens the new .html file in the default browser. Example terminal command line:

./runDiag simple